requirements.txt
requests
python-dotenv

pydub
pyacoustid
mutagen
musicbrainzngs

fuzzywuzzy
python-Levenshtein